backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
apscheduler==3.10.4
requests==2.32.3